+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Req 1.2.0
+
+* Added the `parseUrl` function.
+
 ## Req 1.1.0
 
 * Added `customAuth` and `attachHeader` to facilitate creation of custom